requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
python-telegram-bot==20.8
requests
